+++ tests/testthat/test-check.status.code.R
@@ -1,9 +1,8 @@
-# Copyright (c) 2015-present, Facebook, Inc.
+# Copyright (c) Facebook, Inc. and its affiliates.
 # All rights reserved.
 #
 # This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree. An additional grant
-# of patent rights can be found in the PATENTS file in the same directory.
+# LICENSE file in the root directory of this source tree.
 
 context('check.status.code')
 
@@ -38,6 +37,7 @@ test_that('check.status.code works', {
   )
   expect_error(
     check.status.code(response),
-    '403'
+    '403',
+    class='http_403'
   )
 })