+++ tests/testthat/test-tbl.src_presto.R
@@ -1,9 +1,8 @@
-# Copyright (c) 2015-present, Facebook, Inc.
+# Copyright (c) Facebook, Inc. and its affiliates.
 # All rights reserved.
 #
 # This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree. An additional grant
-# of patent rights can be found in the PATENTS file in the same directory.
+# LICENSE file in the root directory of this source tree.
 
 context('tbl.src_presto')
 
@@ -24,7 +23,21 @@ test_that('dplyr::tbl works', {
     dplyr::sql(paste('SELECT * FROM', parts[['iris_table_name']]))
   )
   expect_equal(
-    iris_presto_materialized,
-    as.data.frame(dplyr::collect(iris_from_sql, n=Inf))
+    dplyr::arrange(
+      iris_presto_materialized,
+      sepal_width,
+      sepal_length,
+      petal_width,
+      petal_length,
+      species
+    ),
+    dplyr::arrange(
+      as.data.frame(dplyr::collect(iris_from_sql, n=Inf)),
+      sepal_width,
+      sepal_length,
+      petal_width,
+      petal_length,
+      species
+    ),
   )
 })