+++ NEWS.md
@@ -1,3 +1,8 @@
+# RPresto 1.3.4
+
+- Translate `[[` to allow indexing arrays and maps with dplyr (#110).
+- Switch from BSD+Patents license to Standard BSD license (#114).
+
 # RPresto 1.3.3
 
 - Fix tests for compatibility with `dbplyr` 1.4.0.